More logging, some exception handling, simplified token management.
