+++ bridge-connector/connector/client/protocol.py
