+++ bridge-connector/setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 
 setup(
     name='bridge-connector',
-    packages=['connector'],
+    packages=['connector', 'connector.client'],
     entry_points={
         'console_scripts': ['bridge-connector=connector.main:main'],
     }
